Restore static export and sync dynamic params
