+++ web/src/app/meriter/communities/[id]/create/page.tsx
@@ -1,7 +1,7 @@
 import { CreatePublicationPageClient } from './CreatePublicationPageClient';
 
 interface CreatePublicationPageProps {
-  params: Promise<{ id: string }>;
+  params: { id: string };
 }
 
 // Required for static export with dynamic routes
@@ -10,8 +10,8 @@ export async function generateStaticParams(): Promise<Array<{ id: string }>> {
   return [];
 }
 
-export default async function CreatePublicationPage({ params }: CreatePublicationPageProps) {
-  const { id } = await params;
+export default function CreatePublicationPage({ params }: CreatePublicationPageProps) {
+  const { id } = params;
   return <CreatePublicationPageClient communityId={id} />;
 }
