+++ web/src/app/meriter/communities/[id]/edit-poll/[pollId]/page.tsx
@@ -1,7 +1,7 @@
 import { EditPollPageClient } from './EditPollPageClient';
 
 interface EditPollPageProps {
-  params: Promise<{ id: string; pollId: string }>;
+  params: { id: string; pollId: string };
 }
 
 // Required for static export with dynamic routes
@@ -10,7 +10,7 @@ export async function generateStaticParams(): Promise<Array<{ id: string; pollId
   return [];
 }
 
-export default async function EditPollPage({ params }: EditPollPageProps) {
-  const { id, pollId } = await params;
+export default function EditPollPage({ params }: EditPollPageProps) {
+  const { id, pollId } = params;
   return <EditPollPageClient communityId={id} pollId={pollId} />;
 }
