+++ web/src/app/meriter/communities/[id]/page.tsx
@@ -1,7 +1,7 @@
 import { CommunityPageClient } from './CommunityPageClient';
 
 interface CommunityPageProps {
-  params: Promise<{ id: string }>;
+  params: { id: string };
 }
 
 // Required for static export with dynamic routes
@@ -10,7 +10,7 @@ export async function generateStaticParams(): Promise<Array<{ id: string }>> {
   return [];
 }
 
-export default async function CommunityPage({ params }: CommunityPageProps) {
-  const { id } = await params;
+export default function CommunityPage({ params }: CommunityPageProps) {
+  const { id } = params;
   return <CommunityPageClient communityId={id} />;
 }
