+++ web/src/app/meriter/communities/[id]/rules/page.tsx
@@ -1,7 +1,7 @@
 import { CommunityRulesPageClient } from './CommunityRulesPageClient';
 
 interface CommunityRulesPageProps {
-  params: Promise<{ id: string }>;
+  params: { id: string };
 }
 
 // Required for static export with dynamic routes
@@ -10,7 +10,7 @@ export async function generateStaticParams(): Promise<Array<{ id: string }>> {
   return [];
 }
 
-export default async function CommunityRulesPage({ params }: CommunityRulesPageProps) {
-  const { id } = await params;
+export default function CommunityRulesPage({ params }: CommunityRulesPageProps) {
+  const { id } = params;
   return <CommunityRulesPageClient communityId={id} />;
 }
